Added a pedantic typesafe integer class, and a configure option (--with-typesafe) to enable it. All coercions between integer types show up as compile errors. Very useful for checking 64-bit integer overflows

+++ src/safe_array.h
@@ -236,7 +236,7 @@ private:
   int64_t len_;
   int32_t rows_;
   int32_t cols_;
-  int32_t resize_chunk_;
+  int64_t resize_chunk_;
   
   T*  data_;
   int64_t data_size_;
